update user and passwod

+++ src/users/users.service.ts
@@ -1,21 +1,26 @@
 import { BadRequestException, Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
-import { Model } from 'mongoose';
+import mongoose, { Model, ObjectId } from 'mongoose';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { CreateUserDto } from './dto/create-user.dto';
 import { User } from './users.model';
+import { updatePassword } from './dto/update-password.dto';
+import * as bycypt from 'bcrypt';
 
+import{AuthService} from '../auth/auth.service'
+import { JwtService } from '@nestjs/jwt';
 @Injectable()
 export class UsersService {
   constructor(
     @InjectModel(User.name) private readonly userModel: Model<User>,
   ) {}
+
   async create(createUserDto: CreateUserDto): Promise<User> {
     try {
       const newUser = await this.userModel.create(createUserDto);
       return newUser;
     } catch (err) {
-      throw new BadRequestException(err);
+      throw new BadRequestException(err.message,err);
     }
   }
 
@@ -30,22 +35,22 @@ export class UsersService {
 
   async findOne(id: string): Promise<User> {
     try {
+      if (!mongoose.Types.ObjectId.isValid(id)){
+        throw new BadRequestException('Provide valid ID')
+      }
       const user = await this.userModel.findById(id);
       if (!user) {
-        throw new BadRequestException(`No user with this email ${id}`);
+        throw new BadRequestException(`No user with this ID : ${id}`);
       }
       return user;
     } catch (err) {
-      throw new BadRequestException(err);
+      throw new BadRequestException(err.message,err);
     }
   }
+
   async findUserByEmail(email: string, selectField?: string): Promise<User> {
     try {
-      let user;
-      if (!selectField) {
-        user = await this.userModel.findOne({ email });
-      }
-      user = await this.userModel.findOne({ email }).select(selectField);
+      const user = await this.userModel.findOne({ email }).select(selectField);
       if (!user) {
         throw new BadRequestException(`No user with this email ${email}`);
       }
@@ -55,27 +60,57 @@ export class UsersService {
     }
   }
 
-  async update(id: string, updateUserDto: UpdateUserDto) {
+  async update(id: string, updateOptions: UpdateUserDto) {
     try {
-      const user = await this.userModel.findByIdAndUpdate(id, updateUserDto);
+      if (!mongoose.Types.ObjectId.isValid(id)){
+        throw new BadRequestException('Provide valid ID')
+      }
+      await this.userModel.updateOne({_id:id}, updateOptions,{runValidators:true});
+      const user =await this.findOne(id)
       if (!user) {
         throw new BadRequestException(`No user with this email ${id}`);
       }
       return user;
     } catch (err) {
-      throw new BadRequestException(err);
+      throw new BadRequestException(err.message,err);
     }
   }
 
-  async remove(id: string) {
-    await this.userModel.findByIdAndDelete(id);
-    return `User with ID ${id} has been deleted`;
+  async updatePassword(id:string,passwordsOpt:updatePassword){
+    try{
+      const user = await this.userModel.findById(id).select('password')
+      if (user && (await bycypt.compare(passwordsOpt.oldPassword, user.password))) {
+      user.password=passwordsOpt.password
+      user.confirmPassword =passwordsOpt.confirmPassword
+      await user.save()
+        return {message:`Your password has been changed please login again with new password`};
+      }
+      throw new BadRequestException('Your old password is wrong!')
+    }catch(err){
+      throw new BadRequestException(err.message,err)
+    }
+    
+  }
+
+  async remove(id: string): Promise<{message:string}> {
+    try {
+      if (!mongoose.Types.ObjectId.isValid(id)) {
+        throw new BadRequestException(`Provide valid ID: ${id}`);
+      }
+      const deletedUser = await this.userModel.findByIdAndDelete(id);
+      if (!deletedUser) {
+        throw new BadRequestException(`No user with this ID : ${id}`);
+      }
+      return { message: `User with ID ${id} has been deleted` };
+    } catch (err) {
+      throw new BadRequestException(err.message, err);
+    }
   }
 
   async giveAdmin(email: string): Promise<User> {
     const user = await this.findUserByEmail(email);
     if (user.roles.includes('admin')) {
-      throw new BadRequestException('User is alredy has admin role');
+      throw new BadRequestException('User is already has admin role');
     }
     user.roles.push('admin');
     user.save({ validateBeforeSave: false });
